requirements.txt
supabase
google-generativeai
pydantic>=2.0
pydantic-ai
rich
python-dotenv
tenacity
